Future date restriction and sorting logs by date added

+++ routers/weights.py
@@ -8,14 +8,21 @@ from sqlalchemy.orm import selectinload
 import models
 from database import get_db
 from schemas import WeightCreate, WeightResponse
+from datetime import date
 
 from auth import CurrentUser
+from validation import date_validation
 
 router = APIRouter()
+
+
+
 
 # add weight
 @router.post("/weight_logs", response_model=WeightResponse, status_code=status.HTTP_201_CREATED)
 async def add_weight(current_user: CurrentUser, weight: WeightCreate, db: Annotated[AsyncSession, Depends(get_db)]):
+
+    date_validation(weight.date)
 
     new_weight = models.Weight(
         weight_kg = weight.weight_kg,
@@ -23,9 +30,41 @@ async def add_weight(current_user: CurrentUser, weight: WeightCreate, db: Annota
         user_id = current_user.id
     )
 
+
+
     db.add(new_weight)
+
+    # This temporarily stores weight in Weight table
+    await db.flush()
+
+    # We select the latest date with highest id
+    result = await db.execute(
+        select(models.Weight)
+        .where(models.Weight.user_id == current_user.id)
+        .order_by(
+            models.Weight.date.desc(),
+            models.Weight.id.desc()
+            )
+        .limit(1)
+    )
+
+    latest_weight = result.scalars().first()
+
+    result = await db.execute(
+        select(models.User)
+        .where(models.User.id == current_user.id)
+    )
+    user = result.scalars().first()
+
+    user.weight_kg = latest_weight.weight_kg
+
+    # commiting both changes at once
+    # This avoids partial updating
     await db.commit()
+
+
     await db.refresh(new_weight)
+
 
     return new_weight
 
@@ -33,7 +72,11 @@ async def add_weight(current_user: CurrentUser, weight: WeightCreate, db: Annota
 @router.get("/weight_logs", response_model=list[WeightResponse])
 async def get_weights_user(current_user: CurrentUser, db: Annotated[AsyncSession, Depends(get_db)]):
 
-    results = await db.execute(select(models.Weight).where(models.Weight.user_id == current_user.id))
+    results = await db.execute(
+        select(models.Weight)
+        .where(models.Weight.user_id == current_user.id)
+        .order_by(models.Weight.date.desc())
+        )
     weights = results.scalars().all()
 
     if weights:
@@ -65,4 +108,32 @@ async def delete_weight(current_user: CurrentUser, weight_id: int, db: Annotated
         )
 
     await db.delete(weight)
+
+    await db.flush()
+
+    # We select the latest date with highest id
+    result = await db.execute(
+        select(models.Weight)
+        .where(models.Weight.user_id == current_user.id)
+        .order_by(
+            models.Weight.date.desc(),
+            models.Weight.id.desc()
+            )
+        .limit(1)
+    )
+
+    latest_weight = result.scalars().first()
+
+    result = await db.execute(
+        select(models.User)
+        .where(models.User.id == current_user.id)
+    )
+    user = result.scalars().first()
+
+    if latest_weight:
+        user.weight_kg = latest_weight.weight_kg
+    else:
+        user.weight_kg = None
+    
+
     await db.commit()
